Initilal Commit
Now I am struggling with file handling.
I am trying to use read and write function for storing the contents of the file. Since read and write function store data in binary format so although I am able to write the contents but unable to retrieve it as the program stops working. 
With the current code I have to re-enter the contents again for each execution of the program